Calculate necessary changes to lithium chloride solutions to achieve target relative humidity

Calculate necessary changes to lithium chloride solutions to achieve target relative humidity

Usage

rh_modify(initial_rh, desired_rh, initial_volume, verbose = TRUE)

Arguments

initial_rh

The relative humidity percentage value of the solution to be modified

desired_rh

The percentage value of the desired relative humidity

initial_volume

The volume 'in Litres' of the initial solution

verbose

Logical. If TRUE, suppresses messages or output during function execution.

Value

a string with instructions to modify the solution

Examples

rh_modify(initial_rh=20,desired_rh=40,initial_volume=0.5)
#> Reminder: Ensure all values are in g. and L.
#>     Add      Quantity
#> 1 water 0.2389975 [L]